Move `appQueryStatusStore` into `runtime-client`
