+++ web-common/src/features/sources/workspace/SourceWorkspaceHeader.svelte
@@ -29,8 +29,8 @@
     V1ReconcileResponse,
     V1Source,
   } from "@rilldata/web-common/runtime-client";
+  import { appQueryStatusStore } from "@rilldata/web-common/runtime-client/application-store";
   import { invalidateAfterReconcile } from "@rilldata/web-common/runtime-client/invalidation";
-  import { appQueryStatusStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
   import { behaviourEvent } from "@rilldata/web-local/lib/metrics/initMetrics";
   import { BehaviourEventMedium } from "@rilldata/web-local/lib/metrics/service/BehaviourEventTypes";
   import {
